test(convertors): split tests into json/table/tree/metrics with shared fixtures
